requirements.txt
Django==1.8
dj-dajaxice==0.8.2
django-dajax==0.9.2
django-dajax-me==0.9.4
django-dajaxice==0.7
django-dajaxice-1.9==0.8.1.1
django-dajaxice-me==0.8
Dajax==1.0
